backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.84.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "groq>=1.1.1",
    "httpx>=0.28.1",
    "huggingface-hub>=1.6.0",
    "opencv-python-headless>=4.13.0.92",
    "pgvector>=0.4.2",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pymupdf>=1.27.2",
    "pymupdf4llm>=1.27.2.1",
    "pytesseract>=0.3.13",
    "python-jose[cryptography]>=3.5.0",
    "python-magic-bin>=0.4.14",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "structlog>=25.5.0",
    "supabase>=2.28.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
lint = [
    "ruff>=0.12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = ["app/models/*", "app/schemas/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
